feat: move icons to icon folder
