+++ src/components/SearchResultSection.tsx
@@ -2,7 +2,7 @@ import type { ResponseDocumentMatch } from "@/lib/common";
 import { cn, getCleanedMetadata } from "@/lib/utils";
 import { useState } from "react";
 import { Link } from "./Link";
-import { AcrobatIcon } from "./ui/acrobat-icon";
+import { AcrobatIcon } from "./ui/icons/acrobat-icon";
 import { GlobeIcon } from "@radix-ui/react-icons";
 
 interface SearchResultProps {
